[REFACTOR] 토큰 타입에 따른 메소드 별 검증

+++ src/orders/orders.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Post, Body, Param, Get, Request, UseGuards, Delete } from '@nestjs/common';
+import { Controller, Post, Body, Param, Get, Request, UseGuards, Delete, UnauthorizedException } from '@nestjs/common';
 import { OrdersService } from './orders.service';
 import { CreateOrderDto } from './dto/create-orders.dto';
 import { Order } from './entities/orders.entity';
@@ -32,15 +32,23 @@ export class OrdersController {
 
   @Delete(':order_id')
   async deleteOrder(@Request() req, @Param('order_id') orderId: string): Promise<void> {
-    const restaurantId = req.user.restaurantId;
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return this.ordersService.deleteOrder(orderId, restaurantId);
+      return this.ordersService.deleteOrder(orderId, restaurantId);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Delete('table/:table_num')
   async deleteOrdersByTableNum(@Request() req, @Param('table_num') tableNum: number): Promise<void> {
-    const restaurantId = req.user.restaurantId;
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return this.ordersService.deleteOrdersByTableNum(restaurantId, tableNum);
+      return this.ordersService.deleteOrdersByTableNum(restaurantId, tableNum);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 }
